carbon0/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
black==20.8b1
boto3==1.14.60
botocore==1.17.60
click==7.1.2
django-allauth==0.42.0
dj-database-url==0.5.0
Django==3.1.1
django-heroku==0.3.1
django-storages==1.10
djangorestframework==3.11.1
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
mypy-extensions==0.4.3
pathspec==0.8.0
Pillow==7.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
regex==2020.7.14
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
whitenoise==5.2.0
